requirements.txt
django==5.1.1
django-graphql-jwt==0.4.0